pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pit"
version = "0.1.0"
description = "PIT - Prompt Injection Tester. Security testing tool for LLM applications."
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2",
    "openai>=1",
    "anthropic>=0.25",
    "pyyaml>=6",
    "questionary>=2.0",
]

[project.scripts]
pit = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["core", "cli"]
